+++ src/Sylius/Bundle/CoreBundle/OAuth/UserProvider.php
@@ -16,7 +16,7 @@ use FOS\UserBundle\Model\UserManagerInterface;
 use HWI\Bundle\OAuthBundle\OAuth\Response\UserResponseInterface;
 use HWI\Bundle\OAuthBundle\Security\Core\User\FOSUBUserProvider;
 use Sylius\Component\Core\Model\UserInterface as SyliusUserInterface;
-use Sylius\Component\Core\Model\UserOAuthInterface;
+use Sylius\Component\User\Model\UserOAuthInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
