+++ src/Sylius/Component/Core/Model/User.php
@@ -12,33 +12,25 @@
 namespace Sylius\Component\Core\Model;
 
 use Doctrine\Common\Collections\ArrayCollection;
-use FOS\UserBundle\Model\User as BaseUser;
+use Sylius\Component\User\Model\User as UserComponent;
 
 /**
  * User model.
  *
  * @author Paweł Jędrzejewski <pawel@sylius.org>
  */
-class User extends BaseUser implements UserInterface
+class User extends UserComponent implements UserInterface
 {
-    protected $firstName;
-    protected $lastName;
-    protected $createdAt;
-    protected $updatedAt;
-    protected $deletedAt;
     protected $currency;
     protected $orders;
     protected $billingAddress;
     protected $shippingAddress;
     protected $addresses;
-    protected $oauthAccounts;
 
     public function __construct()
     {
-        $this->createdAt     = new \DateTime();
         $this->orders        = new ArrayCollection();
         $this->addresses     = new ArrayCollection();
-        $this->oauthAccounts = new ArrayCollection();
 
         parent::__construct();
     }
@@ -153,169 +145,11 @@ class User extends BaseUser implements UserInterface
         return $this->addresses;
     }
 
-    public function getFullName()
-    {
-        return $this->firstName.' '.$this->lastName;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setFirstName($firstName)
-    {
-        $this->firstName = $firstName;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getFirstName()
-    {
-        return $this->firstName;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setLastName($lastName)
-    {
-        $this->lastName = $lastName;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getLastName()
-    {
-        return $this->lastName;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getCreatedAt()
-    {
-        return $this->createdAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setCreatedAt(\DateTime $createdAt)
-    {
-        $this->createdAt = $createdAt;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getUpdatedAt()
-    {
-        return $this->updatedAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setUpdatedAt(\DateTime $updatedAt)
-    {
-        $this->updatedAt = $updatedAt;
-
-        return $this;
-    }
-
     /**
      * {@inheritdoc}
      */
     public function isDeleted()
     {
         return null !== $this->deletedAt && new \DateTime() >= $this->deletedAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDeletedAt()
-    {
-        return $this->deletedAt;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setDeletedAt(\DateTime $deletedAt)
-    {
-        $this->deletedAt = $deletedAt;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setEmail($email)
-    {
-        parent::setEmail($email);
-        parent::setUsername($email);
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setEmailCanonical($emailCanonical)
-    {
-        parent::setEmailCanonical($emailCanonical);
-        parent::setUsernameCanonical($emailCanonical);
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOAuthAccounts()
-    {
-        return $this->oauthAccounts;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOAuthAccount($provider)
-    {
-        if ($this->oauthAccounts->isEmpty()) {
-            return null;
-        }
-
-        $filtered = $this->oauthAccounts->filter(function (UserOAuthInterface $oauth) use ($provider) {
-            return $provider === $oauth->getProvider();
-        });
-
-        if ($filtered->isEmpty()) {
-            return null;
-        }
-
-        return $filtered->current();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function addOAuthAccount(UserOAuthInterface $oauth)
-    {
-        if (!$this->oauthAccounts->contains($oauth)) {
-            $this->oauthAccounts->add($oauth);
-            $oauth->setUser($this);
-        }
-
-        return $this;
     }
 }
