+++ src/Sylius/Component/User/.gitignore
@@ -1,0 +1,5 @@
+vendor/
+bin/
+
+composer.phar
+composer.lock
