+++ src/Sylius/Component/User/.travis.yml
@@ -1,0 +1,13 @@
+language: php
+
+php:
+  - 5.3
+  - 5.4
+  - 5.5
+
+before_script: composer install --dev --prefer-source
+
+script: bin/phpspec run -fpretty --verbose
+
+notifications:
+    irc: "irc.freenode.org#sylius-dev"
