+++ src/Sylius/Component/User/CHANGELOG.md
@@ -1,0 +1,6 @@
+CHANGELOG
+=========
+
+### v0.11.0
+
+* Initial dev release.
