+++ src/Sylius/Component/User/Model/User.php
@@ -1,0 +1,196 @@
+<?php
+
+/*
+* This file is part of the Sylius package.
+*
+* (c) Paweł Jędrzejewski
+*
+* For the full copyright and license information, please view the LICENSE
+* file that was distributed with this source code.
+*/
+
+namespace Sylius\Component\User\Model;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use FOS\UserBundle\Model\User as BaseUser;
+
+/**
+ * User model.
+ *
+ * @author Paweł Jędrzejewski <pawel@sylius.org>
+ */
+class User extends BaseUser implements UserInterface
+{
+    protected $firstName;
+    protected $lastName;
+    protected $oauthAccounts;
+    protected $createdAt;
+    protected $updatedAt;
+    protected $deletedAt;
+    
+    public function __construct()
+    {
+        $this->createdAt     = new \DateTime();
+        $this->oauthAccounts = new ArrayCollection();
+
+        parent::__construct();
+    }
+
+    public function getFullName()
+    {
+        return $this->firstName.' '.$this->lastName;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setFirstName($firstName)
+    {
+        $this->firstName = $firstName;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getFirstName()
+    {
+        return $this->firstName;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setLastName($lastName)
+    {
+        $this->lastName = $lastName;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getLastName()
+    {
+        return $this->lastName;
+    }
+    
+    /**
+     * {@inheritdoc}
+     */
+    public function getOAuthAccounts()
+    {
+        return $this->oauthAccounts;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getOAuthAccount($provider)
+    {
+        if ($this->oauthAccounts->isEmpty()) {
+            return null;
+        }
+
+        $filtered = $this->oauthAccounts->filter(function (UserOAuthInterface $oauth) use ($provider) {
+            return $provider === $oauth->getProvider();
+        });
+
+        if ($filtered->isEmpty()) {
+            return null;
+        }
+
+        return $filtered->current();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function addOAuthAccount(UserOAuthInterface $oauth)
+    {
+        if (!$this->oauthAccounts->contains($oauth)) {
+            $this->oauthAccounts->add($oauth);
+            $oauth->setUser($this);
+        }
+
+        return $this;
+    }
+    
+    /**
+     * {@inheritdoc}
+     */
+    public function getCreatedAt()
+    {
+        return $this->createdAt;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setCreatedAt(\DateTime $createdAt)
+    {
+        $this->createdAt = $createdAt;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getUpdatedAt()
+    {
+        return $this->updatedAt;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setUpdatedAt(\DateTime $updatedAt)
+    {
+        $this->updatedAt = $updatedAt;
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDeletedAt()
+    {
+        return $this->deletedAt;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setDeletedAt(\DateTime $deletedAt)
+    {
+        $this->deletedAt = $deletedAt;
+
+        return $this;
+    }
+    
+    /**
+     * {@inheritdoc}
+     */
+    public function setEmail($email)
+    {
+        parent::setEmail($email);
+        parent::setUsername($email);
+
+        return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setEmailCanonical($emailCanonical)
+    {
+        parent::setEmailCanonical($emailCanonical);
+        parent::setUsernameCanonical($emailCanonical);
+
+        return $this;
+    }
+}
