+++ src/Sylius/Component/User/Model/UserOAuth.php
@@ -9,7 +9,7 @@
 * file that was distributed with this source code.
 */
 
-namespace Sylius\Component\Core\Model;
+namespace Sylius\Component\User\Model;
 
 /**
  * User OAuth model.
