+++ src/Sylius/Component/User/README.md
@@ -1,0 +1,47 @@
+User Component [![Build status...](https://secure.travis-ci.org/Sylius/Locale.png?branch=master)](http://travis-ci.org/Sylius/Locale)
+================
+
+Managing users in PHP applications.
+
+Sylius
+------
+
+Modern ecommerce for PHP and Symfony2. Visit [Sylius.org](http://sylius.org).
+
+Documentation
+-------------
+
+Documentation is available on [**docs.sylius.org**](http://docs.sylius.org/en/latest/components/Users/index.html).
+
+Contributing
+------------
+
+All instructions for contributing to Sylius can be found in the [Contributing Guide](http://docs.sylius.org/en/latest/contributing/index.html).
+
+Support
+-------
+
+If you have a question regarding the usage of this library, please ask on
+[Stackoverflow](http://stackoverflow.com). You should use "sylius"
+tag when posting and make sure to [browse existing questions](http://stackoverflow.com/questions/tagged/sylius).
+
+Sylius on Twitter
+-----------------
+
+[Follow the official Sylius account on Twitter!](http://twitter.com/Sylius).
+
+Bug Tracking
+------------
+
+If you find a bug, please refer to the [Reporting a Bug](http://docs.sylius.org/en/latest/contributing/code/bugs.html) guide.
+
+MIT License
+-----------
+
+License can be found [here](https://github.com/Sylius/Locale/blob/master/LICENSE).
+
+Authors
+-------
+
+The component was originally created by [Paweł Jędrzejewski](http://pjedrzejewski.com).
+See the list of [contributors](https://github.com/Sylius/Locale/contributors).
