+++ src/Sylius/Component/User/composer.json
@@ -1,0 +1,45 @@
+{
+    "name": "sylius/user",
+    "type": "library",
+    "description": "Locale handling for PHP applications.",
+    "keywords": ["language", "localization", "locale", "i18n", "internationalization"],
+    "homepage": "http://sylius.org",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Paweł Jędrzejewski",
+            "homepage": "http://pjedrzejewski.com"
+        },
+        {
+            "name": "Sylius project",
+            "homepage": "http://sylius.org"
+        },
+        {
+            "name": "Community contributions",
+            "homepage": "http://github.com/Sylius/Sylius/contributors"
+        }
+    ],
+    "require": {
+        "php": ">=5.3.3",
+
+        "sylius/resource": "0.13.*@dev",
+        "sylius/storage": "0.13.*@dev"
+    },
+    "require-dev": {
+        "phpspec/phpspec": "~2.1"
+    },
+    "config": {
+        "bin-dir": "bin"
+    },
+    "autoload": {
+        "psr-4": { "Sylius\\Component\\User\\": "" }
+    },
+    "autoload-dev": {
+        "psr-4": { "Sylius\\Component\\User\\spec\\": "spec/" }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "0.13-dev"
+        }
+    }
+}
