+++ src/Sylius/Component/User/phpspec.yml
@@ -1,0 +1,5 @@
+suites:
+    main:
+        namespace: Sylius\Component\User
+        psr4_prefix: Sylius\Component\User
+        src_path: .
